refactor: changed button color and renamed project
