+++ webforj-data-bind/pom.xml
@@ -4,11 +4,11 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.webforj.nestedbeans</groupId>
-  <artifactId>webforj-nestedbeans</artifactId>
+  <groupId>com.webforj.databind</groupId>
+  <artifactId>webforj-databind</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>war</packaging>
-  <name>webforj-nestedbeans</name>
+  <name>webforj-databind</name>
 
   <properties>
     <webforj.version>26.01-SNAPSHOT</webforj.version>
