+++ webforj-data-bind/src/main/java/com/webforj/databind/Application.java
@@ -1,4 +1,4 @@
-package com.webforj.nestedbeans;
+package com.webforj.databind;
 
 import com.webforj.App;
 import com.webforj.annotation.AppProfile;
@@ -6,9 +6,9 @@ import com.webforj.annotation.AppTheme;
 import com.webforj.annotation.Routify;
 import com.webforj.annotation.StyleSheet;
 
-@Routify(packages = "com.webforj.nestedbeans.views")
+@Routify(packages = "com.webforj.databind.views")
 @StyleSheet("ws://app.css")
 @AppTheme("system")
-@AppProfile(name = "webforj-nestedbeans", shortName = "webforj-nestedbeans")
+@AppProfile(name = "webforj-databind", shortName = "webforj-databind")
 public class Application extends App {
 }
