+++ webforj-data-bind/src/main/java/com/webforj/databind/domain/Address.java
@@ -1,4 +1,4 @@
-package com.webforj.nestedbeans.domain;
+package com.webforj.databind.domain;
 
 import jakarta.persistence.Embeddable;
 import jakarta.validation.constraints.NotBlank;
