+++ webforj-data-bind/src/main/resources/webforj.conf
@@ -1,6 +1,6 @@
 # webforJ configuration file (HOCON format).
 # See: https://docs.webforj.com/docs/configuration/properties
 
-webforj.entry = com.webforj.nestedbeans.Application
+webforj.entry = com.webforj.databind.Application
 webforj.debug = true
 webforj.reloadOnServerError = on
